Revert "change package name to org.oneedu."
This reverts commit 348c0b63e3c156e53dc44ac2ff9735e63332e591.
