+++ analytics/src/main/java/org/OneEducation/HarvestClient/HarvestEntry.java
@@ -17,35 +17,26 @@
  * USA
  */
 
-package org.oneedu.HarvestClient;
+package org.OneEducation.HarvestClient;
 
-import android.app.Activity;
-import android.os.Bundle;
-import android.content.Intent;
-import android.view.View;
+import java.lang.Long;
+import java.lang.String;
+import java.lang.System;
 
-public class HarvestActivity extends Activity
-{
-    /** Called when the activity is first created. */
-    @Override
-    public void onCreate(Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.main);
 
-        View okButton = findViewById(R.id.okbutton);
-        okButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                finish();
-            }
-        });
+class HarvestEntry {
 
-        launchService();
+    public String packageName;
+    public Long started;
+    public Long duration;
+
+    HarvestEntry (String _packageName) {
+        packageName = _packageName;
+        started = 0L;
+        duration = 0L;
     }
 
-    private void launchService(){
-        Intent serviceIntent = new Intent(this, HarvestService.class);
-        startService(serviceIntent);
+    public void increment(Long delta) {
+        duration += delta;
     }
 }
