+++ analytics/src/main/java/org/OneEducation/HarvestClient/HarvestSettings.java
@@ -17,7 +17,7 @@
  * USA
  */
 
-package org.oneedu.HarvestClient;
+package org.OneEducation.HarvestClient;
 
 import java.lang.Long;
 import java.lang.System;
